# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.WorkoutRequest:
    properties:
      calories_burned:
        type: integer
      duration_minutes:
        type: integer
      notes:
        type: string
      type:
        type: string
      user_id:
        type: integer
    required:
    - duration_minutes
    - type
    - user_id
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для фитнес-клуба Energy с PostgreSQL
  title: Fitness Club Energy API
  version: "1.0"
paths:
...
  /api/v1/workouts:
    get:
      description: Получение тренировок с фильтрацией по типу, дате, пользователю
      parameters:
      - description: Тип тренировки
        in: query
        name: type
        type: string
      - description: Дата тренировки (YYYY-MM-DD)
        in: query
        name: date
        type: string
      - description: ID пользователя
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: Получить тренировки с динамическими фильтрами
      tags:
      - workouts
    post:
      consumes:
      - application/json
      description: Принимает динамические данные о тренировке от клиента
      parameters:
      - description: Динамические данные тренировки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.WorkoutRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Создать тренировку с динамическими данными
      tags:
      - workouts
swagger: "2.0"